just moved an api call from location to tour data file. updated imports where used
