+++ src/api/mergedData.js
@@ -1,9 +1,10 @@
-import { getLocationTours, getSingleLocationById } from './locationData';
+import { getSingleLocation } from './locationData';
+import getLocationTours from './tourData';
 
 // below is a little different from our wireframe because it gives us access to the tours at the location. keeping for now in case we might want to show the tours offered at a location in a location's details page
 const viewLocationDetails = (LocationId) =>
   new Promise((resolve, reject) => {
-    Promise.all([getSingleLocationById(LocationId), getLocationTours(LocationId)])
+    Promise.all([getSingleLocation(LocationId), getLocationTours(LocationId)])
       .then(([locationObject, locationToursArray]) => {
         resolve({ ...locationObject.fields, tours: locationToursArray });
       })
